requirements.txt
PyYAML==6.0
open3d==0.16.0
numpy==1.24.3
opencv-python==4.8.0
pytest==7.4.0
